minor: move gsekit query
