+++ pipeline_plugins/components/query/sites/open/query.py
@@ -13,12 +13,14 @@ specific language governing permissions and limitations under the License.
 
 from pipeline_plugins.components.query.sites.open.cc import cc_urlpatterns
 from pipeline_plugins.components.query.sites.open.file_upload import file_upload_urlpatterns
+from pipeline_plugins.components.query.sites.open.gsekit import gsekit_urlpatterns
 from pipeline_plugins.components.query.sites.open.job import job_urlpatterns
 from pipeline_plugins.components.query.sites.open.nodeman import nodeman_urlpatterns
 from pipeline_plugins.components.query.sites.open.monitor import monitor_urlpatterns
 
 urlpatterns = cc_urlpatterns
 urlpatterns += file_upload_urlpatterns
+urlpatterns += gsekit_urlpatterns
 urlpatterns += job_urlpatterns
 urlpatterns += nodeman_urlpatterns
 urlpatterns += monitor_urlpatterns
